+++ src/game/chest.rs
@@ -30,12 +30,12 @@ impl Chest {
         let distance = &game.location.distance_from_home();
         let gold_chest = random().gold_chest(distance);
         let equipment_chest = random().equipment_chest(distance);
-        let ring_chest = random().ring_chest(distance);
+        let mut ring_chest = random().ring_chest(distance);
 
         let mut chest = Self::default();
 
         if gold_chest {
-            chest.gold = random().gold_gained(game.player.level * 200)
+            chest.gold = random().gold_gained(game.player.level * 50)
         }
 
         if equipment_chest {
@@ -51,6 +51,9 @@ impl Chest {
             // be included in the chest
             if let Some(ring) = random_ring(game) {
                 chest.items.push(Box::new(ring));
+            } else {
+                // only show chest found if there are rings left to be found
+                ring_chest = false;
             }
         }
 
@@ -65,7 +68,7 @@ impl Chest {
         }
 
         // Return None instead of an empty chest if none was found
-        if gold_chest || equipment_chest || item_chest {
+        if gold_chest || equipment_chest || item_chest || ring_chest {
             Some(chest)
         } else {
             None